test: distribute commands in isolation
Adds unit tests for the distributeCommand methods of the
CommandDistributionBehavior class.

I've explicitly chosen not to test this by adding special deployment
tests, because I don't think that makes sense. This behavior can be used
for different commands, not just deployment records.

Instead of spinning up a test engine, this is a must more focused test
that verifies that records are added to the processing result builder
and that there are interactions with the InterPartitionCommandSender.

To support this, I've introduced a new FakeProcessingResultBuilder that
can be used in such tests.